firestore/video-collection/requirements.txt
google-cloud-firestore
google-cloud-pubsub